# shared-api/src/main/resources/openapi/canineai-spec.yaml
openapi: 3.0.3
info:
  title: CanineAI REST API
  description: API contract between the core Backend REST service, the Web Application, the Android Application, and the AI Pipeline Service.
  version: 1.0.0
servers:
  - url: /api/v1
    description: Standard gateway base URL path
paths:
...
  /patients:
    get:
      summary: Get paginated list of patients
      tags:
        - Patients
      parameters:
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          required: false
          schema:
            type: integer
            default: 10
        - name: sort
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Paginated patients retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedPatientsResponse'
    post:
      summary: Admit a new patient
      tags:
        - Patients
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatientRequest'
      responses:
        '200':
          description: Patient admitted successfully
...
components:
  schemas:
...
    PatientRequest:
      type: object
      required:
        - fullName
        - dateOfBirth
        - gender
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: string
          format: date
        gender:
          type: string
    PagedPatientsResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/PatientRequest'
        totalPages:
          type: integer
        totalElements:
          type: integer
...